refactor: 주석삭제 및 추가(#356)
