+++ module-api/src/main/java/com/back2basics/domain/projectstep/controller/ProjectStepController.java
@@ -1,6 +1,7 @@
 package com.back2basics.domain.projectstep.controller;
 
 import static com.back2basics.domain.projectstep.controller.code.ProjectStepResponseCode.*;
+
 import com.back2basics.domain.projectstep.dto.request.CreateProjectStepRequest;
 import com.back2basics.domain.projectstep.dto.request.UpdateProjectStepRequest;
 import com.back2basics.domain.projectstep.dto.response.DetailProjectStepResponse;
@@ -12,7 +13,7 @@ import com.back2basics.projectstep.port.in.DeleteProjectStepUseCase;
 import com.back2basics.projectstep.port.in.ReadProjectStepUseCase;
 import com.back2basics.projectstep.port.in.command.CreateProjectStepCommand;
 import com.back2basics.projectstep.port.in.command.UpdateProjectStepCommand;
-import com.back2basics.projectstep.port.in.command.UpdateProjectStepUseCase;
+import com.back2basics.projectstep.port.in.UpdateProjectStepUseCase;
 import com.back2basics.projectstep.service.result.DetailProjectStepResult;
 import com.back2basics.projectstep.service.result.ReadProjectStepResult;
 import java.util.List;
@@ -30,7 +31,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/api/step")
+@RequestMapping("/api/projects/steps")
 public class ProjectStepController {
 
     private final CreateProjectStepUseCase createProjectStepUseCase;
@@ -38,32 +39,36 @@ public class ProjectStepController {
     private final DeleteProjectStepUseCase deleteProjectStepUseCase;
     private final ReadProjectStepUseCase readProjectStepUseCase;
 
-    // /api/step?projectId={projectId}
+    // /api/projects/steps?projectId={projectId}
     @PostMapping
     public ResponseEntity<ApiResponse<Void>> createStep(
         @RequestBody CreateProjectStepRequest request, @RequestParam Long projectId) {
-        System.out.println("stepOrder in request DTO = " + request.stepOrder()); // 이거 하니까 왜 되는거임 ;;
         CreateProjectStepCommand command = request.toCommand();
         createProjectStepUseCase.createStep(command, projectId);
         return ApiResponse.success(STEP_CREATE_SUCCESS);
     }
 
+    // 프로젝트 별 단계 조회
     @GetMapping("/{projectId}")
     public ResponseEntity<ApiResponse<List<DetailProjectStepResponse>>> getStepsProjectId(
         @PathVariable Long projectId) {
-        List<DetailProjectStepResult> result = readProjectStepUseCase.findDetailByProjectId(projectId);
+        List<DetailProjectStepResult> result = readProjectStepUseCase.findDetailByProjectId(
+            projectId);
         List<DetailProjectStepResponse> response = result.stream()
             .map(DetailProjectStepResponse::toResponse).toList();
         return ApiResponse.success(STEP_ALL_READ_SUCCESS, response);
     }
 
+    // 단계 상세 조회
     @GetMapping("/{stepId}")
-    public ResponseEntity<ApiResponse<ReadProjectStepResponse>> getStepById(@PathVariable Long stepId) {
+    public ResponseEntity<ApiResponse<ReadProjectStepResponse>> getStepById(
+        @PathVariable Long stepId) {
         ReadProjectStepResult result = readProjectStepUseCase.findById(stepId);
         ReadProjectStepResponse response = ReadProjectStepResponse.toResponse(result);
         return ApiResponse.success(STEP_READ_SUCCESS, response);
     }
 
+    // 수정
     @PutMapping("/{stepId}")
     public ResponseEntity<ApiResponse<Void>> updateStep(@PathVariable Long stepId, @RequestBody
     UpdateProjectStepRequest request) {
@@ -81,6 +86,7 @@ public class ProjectStepController {
         return ApiResponse.success(STEP_UPDATE_SUCCESS);
     }
 
+    // 삭제
     @DeleteMapping("/{stepId}")
     public ResponseEntity<ApiResponse<Void>> deleteStep(@PathVariable Long stepId) {
         deleteProjectStepUseCase.softDelete(stepId);
