--- module-api/src/main/java/com/back2basics/domain/projectstep/dto/request/UpdateApprovalStatusRequest.java
@@ -1,9 +1,0 @@
-package com.back2basics.domain.projectstep.dto.request;
-
-import com.back2basics.projectstep.model.ProjectFeedbackStepStatus;
-import jakarta.validation.constraints.NotNull;
-
-public record UpdateApprovalStatusRequest(
-    @NotNull ProjectFeedbackStepStatus projectFeedbackStepStatus) {
-
-}
